requirements.txt
python-docx>=0.8.11
pyyaml>=5.4.1
requests>=2.0.0
beautifulsoup4>=4.0.0
